refactor: fix 4 architectural violations (maintainability)
Task 1 — Extract report parser from action_provider_fs.py
Move 19 standalone parsing/utility functions, RunInfo dataclass, and
grade-order constants to a new adapters/fs/report_parser.py module.
action_provider_fs.py now imports them instead of defining them.

Task 2 — Decouple runner.py and discipline_detector.py from concrete layers
EvaluateConfig gains an optional provider field; run() uses it when set
and lazy-imports FilesystemEvaluators/PracticesRepository only as a
fallback. cli.py constructs and injects the provider. detect_discipline()
and detect_from_rules() accept an optional registry parameter so callers
(and tests) can inject one without touching default_paths().

Task 3 — Remove file I/O from build_scoring_prompt
build_scoring_prompt() now accepts evidence_content/scores_content strings
instead of file paths. File reads moved to evaluation.py so the prompt
builder is pure and testable without touching disk.

Task 4 — CLI normalization and validation cleanup
cli.py evaluate branch delegates to parse_cli_args() (single source of
truth) and now correctly passes evidence_only/no_prescan/numerical.
Empty-string guard for --generate-maps moved from config/cli.py dispatch
to run_generate_evaluators() in config/actions.py where it belongs.
