+++ src/codecompass/action_provider_fs.py
@@ -1,455 +1,37 @@
 from __future__ import annotations
 
-from dataclasses import dataclass
-from datetime import datetime
 import json
 import os
 from pathlib import Path
-import re
 import shutil
 import sys
 from typing import Any
-import urllib.request
 
 from codecompass.action_provider import ActionProvider
 from codecompass.action_provider_jobs import JobManager
+from codecompass.adapters.fs.report_parser import (
+    RunInfo,
+    _build_repository_info,
+    _build_totals,
+    _calculate_trend,
+    _clean_cell,
+    _extract_exec_summary,
+    _get_previous_run_for_dimension,
+    _is_divider_row,
+    _list_runs,
+    _most_frequent_grade,
+    _parse_eval_from_json,
+    _parse_eval_markdown,
+    _parse_evidence_file,
+    _parse_numeric_score,
+    _parse_report_json,
+    _parse_run_id_date,
+    _read_run_data,
+    _safe_read_dir,
+    _split_table_row,
+    _summarize_dimensions,
+)
 from codecompass.evaluate.lib.repo_handler import is_repo_url
-
-
-NUMERIC_GRADE_ORDER = ["Critical", "Poor", "Adequate", "Good", "Exemplary"]
-TEXT_GRADE_ORDER = ["Insufficient", "Developing", "Proficient", "Exemplary"]
-SEVERITIES = {"critical", "major", "minor", "unknown"}
-
-
-@dataclass(frozen=True)
-class RunInfo:
-    run_id: str
-    date_iso: str | None
-    date_label: str
-
-
-def _safe_read_dir(path: Path) -> list[os.DirEntry[str]]:
-    try:
-        with os.scandir(path) as it:
-            return list(it)
-    except OSError:
-        return []
-
-
-def _parse_run_id_date(run_id: str) -> tuple[str | None, str]:
-    match = None
-    if len(run_id) >= 8 and run_id[:8].isdigit():
-        match = run_id[:8]
-    if not match:
-        return None, run_id
-    try:
-        parsed = datetime.strptime(match, "%Y%m%d")
-    except ValueError:
-        return None, run_id
-    return parsed.date().isoformat(), parsed.strftime("%b %d, %Y")
-
-
-def _parse_numeric_score(score_text: str | None) -> float | None:
-    if not score_text:
-        return None
-    match = re.search(r"(\d+(?:\.\d+)?)", str(score_text))
-    if not match:
-        return None
-    return float(match.group(1))
-
-
-def _clean_cell(value: str) -> str:
-    return value.replace("**", "").replace("`", "").strip()
-
-
-def _build_repository_info(repo: str, discipline: str | None) -> dict[str, str | None]:
-    if is_repo_url(repo):
-        name = repo.split("/")[-1].replace(".git", "")
-        return {
-            "name": name,
-            "discipline": discipline,
-            "location": "online",
-            "path": repo,
-        }
-    resolved = Path(repo).resolve()
-    return {
-        "name": resolved.name,
-        "discipline": discipline,
-        "location": "local",
-        "path": str(resolved),
-    }
-
-
-def _split_table_row(line: str) -> list[str]:
-    raw = line.strip()
-    no_outer = raw.lstrip("|").rstrip("|")
-    return [_clean_cell(cell) for cell in no_outer.split("|")]
-
-
-def _is_divider_row(line: str) -> bool:
-    return re.match(r"^\s*\|?\s*[-:]+(\s*\|\s*[-:]+)+\s*\|?\s*$", line) is not None
-
-
-def _extract_exec_summary(markdown: str) -> list[str]:
-    lines = markdown.splitlines()
-    start = -1
-    for idx, line in enumerate(lines):
-        if line.strip().lower() == "## executive summary":
-            start = idx
-            break
-    if start < 0:
-        return []
-    result = []
-    for line in lines[start + 1 :]:
-        if line.strip().startswith("## "):
-            break
-        if "|" in line:
-            result.append(line)
-    return result
-
-
-def _parse_eval_markdown(markdown: str, project: str, run_id: str, dimension: str) -> dict[str, Any]:
-    table_lines = [line for line in _extract_exec_summary(markdown) if not _is_divider_row(line)]
-    principle_grades: list[dict[str, Any]] = []
-    if len(table_lines) >= 2:
-        header_cells = [c for c in _split_table_row(table_lines[0]) if c]
-        is_four_col = len(header_cells) >= 4
-        for line in table_lines[1:]:
-            cells = [c for c in _split_table_row(line) if c]
-            if len(cells) < 2:
-                continue
-            principle = cells[0]
-            score = None
-            grade = None
-            if is_four_col:
-                raw = cells[-1]
-                match = re.match(r"^(\d+(?:\.\d+)?/10)(?:\s+(\w+))?$", raw)
-                if match:
-                    score = match.group(1)
-                    grade = match.group(2)
-                else:
-                    score = raw
-            elif len(cells) >= 3:
-                score = cells[1]
-                grade = cells[2]
-            else:
-                grade = cells[1]
-            if not grade and score:
-                grade_score = _parse_numeric_score(score)
-                if grade_score is not None:
-                    if grade_score >= 9:
-                        grade = "Exemplary"
-                    elif grade_score >= 7:
-                        grade = "Good"
-                    elif grade_score >= 5:
-                        grade = "Adequate"
-                    elif grade_score >= 3:
-                        grade = "Poor"
-                    else:
-                        grade = "Critical"
-            is_overall = "overall" in principle.lower()
-            principle_grades.append(
-                {
-                    "principle": principle,
-                    "score": score,
-                    "grade": grade,
-                    "isOverall": is_overall,
-                }
-            )
-
-    return {
-        "dimension": dimension,
-        "runId": run_id,
-        "project": project,
-        "principleGrades": principle_grades,
-        "principles": [],
-        "priorityRemediation": {"critical": [], "major": [], "minor": []},
-        "rawContent": markdown,
-    }
-
-
-def _most_frequent_grade(grades: list[str]) -> str | None:
-    if not grades:
-        return None
-    counts: dict[str, int] = {}
-    for grade in grades:
-        counts[grade] = counts.get(grade, 0) + 1
-    winner = grades[0]
-    winner_count = counts[winner]
-    for grade, count in counts.items():
-        if count > winner_count:
-            winner = grade
-            winner_count = count
-            continue
-        if count == winner_count:
-            if grade in NUMERIC_GRADE_ORDER and winner in NUMERIC_GRADE_ORDER:
-                if NUMERIC_GRADE_ORDER.index(grade) > NUMERIC_GRADE_ORDER.index(winner):
-                    winner = grade
-                    continue
-            if grade in TEXT_GRADE_ORDER and winner in TEXT_GRADE_ORDER:
-                if TEXT_GRADE_ORDER.index(grade) > TEXT_GRADE_ORDER.index(winner):
-                    winner = grade
-    return winner
-
-
-def _build_totals(violations: list[dict[str, Any]], compliance: list[dict[str, Any]]) -> dict[str, Any]:
-    severity = {"critical": 0, "major": 0, "minor": 0, "unknown": 0}
-    for entry in violations:
-        key = entry.get("severity", "unknown")
-        if key not in SEVERITIES:
-            key = "unknown"
-        severity[key] += 1
-    return {
-        "violationCount": len(violations),
-        "complianceCount": len(compliance),
-        "severity": severity,
-    }
-
-
-def _parse_report_json(json_path: Path) -> dict[str, Any] | None:
-    try:
-        data = json.loads(json_path.read_text())
-    except OSError:
-        return None
-    except json.JSONDecodeError:
-        return None
-
-    violations = [
-        {
-            "principle": v.get("principle"),
-            "file": v.get("file"),
-            "line": v.get("line"),
-            "reason": v.get("reason"),
-            "severity": v.get("severity", "minor"),
-            "snippet": v.get("snippet"),
-        }
-        for v in data.get("violations", [])
-    ]
-    compliance = [
-        {
-            "principle": c.get("principle"),
-            "file": c.get("file"),
-            "line": c.get("line"),
-            "reason": c.get("reason"),
-            "snippet": c.get("snippet"),
-        }
-        for c in data.get("compliance", [])
-    ]
-
-    return {
-        "dimension": data.get("dimension"),
-        "overallScore": data.get("overallScore"),
-        "overallGrade": data.get("overallGrade"),
-        "principles": [
-            {"name": p.get("name"), "score": p.get("score"), "grade": p.get("grade")}
-            for p in data.get("principles", [])
-        ],
-        "detailPrinciples": [],
-        "violations": violations,
-        "compliance": compliance,
-        "totals": _build_totals(violations, compliance),
-    }
-
-
-def _parse_evidence_file(evidence_path: Path) -> dict[str, Any]:
-    dimension = evidence_path.name.replace("_evidence.json", "")
-    try:
-        data = json.loads(evidence_path.read_text())
-    except OSError:
-        data = {}
-    except json.JSONDecodeError:
-        data = {}
-    return {
-        "dimension": dimension,
-        "sourceFileCount": data.get("source_file_count"),
-        "date": data.get("date"),
-        "discipline": data.get("discipline"),
-    }
-
-
-def _summarize_dimensions(dimensions: list[dict[str, Any]]) -> dict[str, Any]:
-    overall_grades = [d.get("overallGrade") for d in dimensions if d.get("overallGrade")]
-    numeric_scores = [
-        score for score in (_parse_numeric_score(d.get("overallScore")) for d in dimensions) if score is not None
-    ]
-    numeric_average = None
-    if numeric_scores:
-        numeric_average = round(sum(numeric_scores) / len(numeric_scores), 1)
-
-    grade_breakdown: dict[str, int] = {}
-    for grade in overall_grades:
-        grade_breakdown[grade] = grade_breakdown.get(grade, 0) + 1
-
-    return {
-        "dimensionsCount": len(dimensions),
-        "overallGrade": _most_frequent_grade(overall_grades),
-        "numericAverage": numeric_average,
-        "gradeBreakdown": [
-            {"grade": grade, "count": count}
-            for grade, count in sorted(grade_breakdown.items(), key=lambda item: (-item[1], item[0]))
-        ],
-    }
-
-
-def _read_run_data(reports_root: Path, project: str, run_id: str) -> list[dict[str, Any]]:
-    run_dir = reports_root / project / run_id
-    evaluation_dir = run_dir / "evaluation"
-    evidence_dir = run_dir / "evidence"
-
-    evaluations: list[dict[str, Any]] = []
-    seen_dimensions: set[str] = set()
-    entries = _safe_read_dir(evaluation_dir)
-    for entry in entries:
-        if not entry.is_file() or not entry.name.endswith("_eval.md"):
-            continue
-        dimension = entry.name[:-8]
-        json_path = evaluation_dir / f"{dimension}.json"
-        parsed = _parse_report_json(json_path) if json_path.exists() else None
-        if parsed:
-            evaluations.append(parsed)
-            seen_dimensions.add(dimension)
-
-    for entry in entries:
-        if not entry.is_file() or not entry.name.endswith(".json"):
-            continue
-        dimension = entry.name[:-5]
-        if dimension in seen_dimensions:
-            continue
-        parsed = _parse_report_json(Path(entry.path))
-        if parsed:
-            evaluations.append(parsed)
-            seen_dimensions.add(dimension)
-
-    evidence_map: dict[str, dict[str, Any]] = {}
-    for entry in _safe_read_dir(evidence_dir):
-        if entry.is_file() and entry.name.endswith("_evidence.json"):
-            parsed = _parse_evidence_file(Path(entry.path))
-            evidence_map[parsed["dimension"]] = parsed
-
-    dimensions = []
-    for evaluation in evaluations:
-        dimension = evaluation.get("dimension")
-        evidence = evidence_map.get(dimension, {})
-        dimensions.append(
-            {
-                **evaluation,
-                "sourceFileCount": evidence.get("sourceFileCount"),
-                "evidenceDate": evidence.get("date"),
-                "discipline": evidence.get("discipline"),
-            }
-        )
-
-    dimensions.sort(key=lambda item: item.get("dimension") or "")
-    return dimensions
-
-
-def _list_runs(reports_root: Path, project: str) -> list[RunInfo]:
-    project_dir = reports_root / project
-    run_infos: list[RunInfo] = []
-    for entry in _safe_read_dir(project_dir):
-        if not entry.is_dir() or entry.name.startswith("."):
-            continue
-        date_iso, date_label = _parse_run_id_date(entry.name)
-        run_infos.append(RunInfo(run_id=entry.name, date_iso=date_iso, date_label=date_label))
-    run_infos.sort(key=lambda r: (r.date_iso or "", r.run_id), reverse=True)
-    return run_infos
-
-
-def _calculate_trend(current_score: Any, previous_score: Any) -> str:
-    current = _parse_numeric_score(str(current_score)) if current_score is not None else None
-    previous = _parse_numeric_score(str(previous_score)) if previous_score is not None else None
-    if current is None or previous is None:
-        return "none"
-    if current > previous:
-        return "up"
-    if current < previous:
-        return "down"
-    return "same"
-
-
-def _get_previous_run_for_dimension(reports_root: Path, project: str, current_run_id: str, dimension: str) -> dict[str, Any] | None:
-    project_path = reports_root / project
-    if not project_path.exists():
-        return None
-    runs = [entry.name for entry in _safe_read_dir(project_path) if entry.is_dir() and entry.name < current_run_id]
-    runs.sort(reverse=True)
-    for run_id in runs:
-        dimensions = _read_run_data(reports_root, project, run_id)
-        dim = next((d for d in dimensions if d.get("dimension") == dimension), None)
-        if dim:
-            return {"runId": run_id, "dimension": dim}
-    return None
-
-
-def _parse_eval_from_json(json_path: Path, project: str, run_id: str, dimension: str) -> dict[str, Any] | None:
-    try:
-        data = json.loads(json_path.read_text())
-    except OSError:
-        return None
-    except json.JSONDecodeError:
-        return None
-
-    principle_grades = [
-        {
-            "principle": p.get("name"),
-            "score": p.get("score"),
-            "grade": p.get("grade"),
-            "isOverall": False,
-        }
-        for p in data.get("principles", [])
-    ]
-    principle_grades.append(
-        {
-            "principle": "Overall",
-            "score": data.get("overallScore"),
-            "grade": data.get("overallGrade"),
-            "isOverall": True,
-        }
-    )
-
-    # Build per-principle detail by grouping top-level violations and compliance
-    principle_map: dict[str, Any] = {}
-    for p in data.get("principles", []):
-        name = p.get("name", "")
-        principle_map[name] = {
-            "name": name,
-            "score": p.get("score"),
-            "grade": p.get("grade"),
-            "violations": [],
-            "compliance": [],
-            "justification": "",
-            "recommendations": [],
-            "metrics": None,
-        }
-    for v in data.get("violations", []):
-        key = v.get("principle", "")
-        if key not in principle_map:
-            principle_map[key] = {"name": key, "score": None, "grade": None, "violations": [], "compliance": [], "justification": "", "recommendations": [], "metrics": None}
-        f = v.get("file")
-        line = v.get("line")
-        principle_map[key]["violations"].append({
-            "code": v.get("snippet", ""),
-            "severity": v.get("severity", "minor"),
-            "file": f"{f}:{line}" if f and line else f,
-            "reason": v.get("reason", ""),
-        })
-    for c in data.get("compliance", []):
-        key = c.get("principle", "")
-        if key not in principle_map:
-            principle_map[key] = {"name": key, "score": None, "grade": None, "violations": [], "compliance": [], "justification": "", "recommendations": [], "metrics": None}
-        principle_map[key]["compliance"].append(c.get("snippet") or c.get("reason") or "")
-
-    return {
-        "dimension": dimension,
-        "runId": run_id,
-        "project": project,
-        "principleGrades": principle_grades,
-        "principles": list(principle_map.values()),
-        "violations": data.get("violations", []),
-        "compliance": data.get("compliance", []),
-        "priorityRemediation": {"critical": [], "major": [], "minor": []},
-        "rawContent": None,
-    }
 
 
 class FilesystemActionProvider(ActionProvider):
@@ -818,6 +400,8 @@ class FilesystemActionProvider(ActionProvider):
         return {"models": models}
 
     def _get_claude_models(self):
+        import json
+        import urllib.request
         # Try direct API call when ANTHROPIC_API_KEY is available
         api_key = os.environ.get("ANTHROPIC_API_KEY")
         if api_key:
